Test

See merge request bms/scmp/scmp-api!492

+++ src/main/java/com/aiqin/bms/scmp/api/purchase/domain/pojo/transport/TransportOrders.java
@@ -26,7 +26,7 @@ public class TransportOrders extends CommonBean {
 
     @ApiModelProperty("商品金额")
 //    @JsonProperty("commodity_amount")
-    private Long commodityAmount;
+    private BigDecimal commodityAmount;
 
     @ApiModelProperty("订单金额")
 //    @JsonProperty("order_amount")
